+++ cbits/img_ffi.h
@@ -1,0 +1,7 @@
+#ifndef IMG_FFI_H_
+#define IMG_FFI_H_
+
+#include <SDL2/SDL.h>
+#include <SDL2/SDL_image.h>
+#endif // IMG_FFI_H_
+