+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for fort
 
+## 0.1.3.0 -- 2024-02-21
+    added append-build to primitives
+    fixed where clause bug
+    started sdl, ttf, img bindings in lib
+
 ## 0.1.2.0 -- 2024-02-15
     created fort standard library in lib
     fixed sum alt default value bug (delete matched tags)